requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
fonttools==4.60.0
frozenlist==1.7.0
gitdb==4.0.12
GitPython==3.1.45
greenlet==3.2.4
groq==0.31.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kagglehub==0.3.13
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-groq==0.3.8
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.30
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.5.0
numpy==2.3.3
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pandas==2.3.2
pillow==11.3.0
propcache==0.3.2
protobuf==6.32.1
pyarrow==21.0.0
pydantic==2.11.9
pydantic-settings==2.11.0
pydantic_core==2.33.2
pydeck==0.9.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.43
streamlit==1.50.0
tenacity==9.1.2
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.25.0
